arrumando os DTOs e APIService

+++ app/src/main/java/com/example/medicoapplication/data/remote/DTO/PageDto.kt
@@ -1,8 +1,8 @@
 package com.example.medicoapplication.data.remote.DTO
 
 data class PageDto(
-    val size: Int,
-    val totalElements: Int,
-    val totalPages: Int,
-    val number: Int
+    val size: Long,
+    val totalElements: Long,
+    val totalPages: Long,
+    val number: Long
 )
